add AccessConfiguration + move dto

+++ scrapper/src/main/java/edu/java/scrapper/repository/jpa/JpaLinkRepository.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JpaLinkRepository extends JpaRepository<Link, Long> {
+public interface JpaLinkRepository extends JpaRepository<Link, Integer> {
     @NotNull
     Optional<Link> findByUrl(String url);
 }
